requirements.txt
streamlit
pandas
scikit-learn
joblib
matplotlib
seaborn
openpyxl
Pillow